refactor service user profiles module
