+++ src/service-user-profiles/service-user-profiles.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { UserEntity } from 'src/entities/user.entity';
+import { UserService } from 'src/user/user.service';
+import { ServiceUserProfilesService } from './service-user-profiles.service';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([UserEntity])],
+  providers: [ServiceUserProfilesService, UserService],
+})
+export class ServiceUserProfilesModule {}
