+++ src/user/user.controller.ts
@@ -104,12 +104,4 @@ export class UserController {
     const users = await this.userService.getUsers(userQuery, include || [], fields, limit);
     return users.map((u) => formatUserObject(u));
   }
-
-  // Use only if users have not been added to mailchimp due to e.g. an ongoing bug
-  @ApiBearerAuth()
-  @Post('/bulk-mailchimp-upload')
-  @UseGuards(FirebaseAuthGuard)
-  async bulkUploadMailchimpProfiles() {
-    return await this.userService.bulkUploadMailchimpProfiles();
-  }
 }
